chore: refactor popup related content utilities
